requirements.txt
anthropic>=0.18.0
httpx>=0.24.0,<0.28.0
pytest==8.3.3
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-xdist>=3.5.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.0

# Vector Database Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
